update test names for consistency
