+++ spec/features/admin/admin_can_log_in_and_see_admin_dashboard_spec.rb
