+++ spec/features/admin/admin_can_see_all_orders_spec.rb
