+++ spec/features/user/user_can_place_order_spec.rb
