+++ spec/features/user/user_can_see_cart_contents_spec.rb
