+++ spec/features/user/user_can_view_retired_item_show_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-feature "user visits retired item" do
+feature "User can view retired item" do
   scenario "they visit the item show page" do
     donation = create(:donation, active: false)
