+++ spec/features/visitor/visitor_can_browse_by_category_spec.rb
