+++ spec/features/visitor/visitor_can_see_all_items_spec.rb
